requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
gensim==4.0.1
google-play-scraper==0.1.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
nltk==3.5
numpy==1.20.2
pandas==1.2.3
Pillow==8.2.0
pyemd==0.5.1
pyenchant==3.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
regex==2021.3.17
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
smart-open==5.0.0
threadpoolctl==2.1.0
tqdm==4.59.0
urllib3==1.26.4
vaderSentiment==3.3.2
